update go code
